chore(license): reuse compliance
This commit adds SPDX Licenses and Copyright for every file in the repository.
I tried to find out as best as I could to find the right Copyright holder for every previously existing file.
For autogenerated files there is no copyright holder and the license is CC0-1.0.
For non code files CC-BY-4.0 is used as specified by AMOS.
For code files MIT is used as specified by AMOS.
For external code (gradle wrapper as example) their original license is used.
As Catchall the original LICENSE file from AMOS is kept.

The reuse cli is added to the nix shell, `reuse lint` checks whether the project is compliant.
Signed-off-by: Felix Hilgers <felix.hilgers@fau.de>