ci: Use a self hosted runner
This change speeds up our current CI pipeline by a lot (20min -> 2min).

We are using a self hosted S3 cache on hetzner and a single large
server, also hosted on S3.

The main reason our old CI system was slow is that we have a large
amount of toolchains that are not installed by default on the github
runners. That means that every job in the CI has to download about 10Gb
of data, just for the toolchains.

The new system is way faster, we have a single server with nix
installed. On this server there are multiple github runners. Each of
them share a single nix store. They also share the gradle caches in the
home directory.

The build artifacts are cached to S3 which is both local and cheap.

Signed-off-by: Felix Hilgers <felix.hilgers@fau.de>