feat: add configuration and visualization screens
This commit adds the configuration screen, which for now, is just
filled with mock values. It also adds the visualization screen, which
uses real data, however the selection is using hardcoded values as
currently, there is only one metric. However, a basic design
for later integration with other components is established.

Signed-off-by: Luca Bretting <luca.bretting@fau.de>