refactor(rust): update folder structure
This commit updates the folder structure of the backend and removes deprecated parts.
The old example is removed and now our daemon, ebpf and common code live all in the backend directory.
The protobuf files are moved into the shared crate for both client and backend.
The xtask scripts are updated with the new paths.

Signed-off-by: Felix Hilgers <felix.hilgers@fau.de>
Co-authored-by: Robin Seidl <68117355+mr-kanister@users.noreply.github.com>
Co-authored-by: Tom Weisshuhn <tom.weisshuhn@fau.de>
Co-authored-by: Franz Schlicht <franz.schlicht@gmail.com>
Co-authored-by: Benedikt Zinn <benedikt.wh.zinn@gmail.com>