feat: connect to backend, add processes list, add error handling, add mock client
This commit connects the client to the backend for the configuration screen
and the newly added process screen. It also two variants of the app, one
connecting to the real backend and on to a mocked backend for easy testing.
These variants have seperate Gradle tasks. The processes screen should be
combined with the configuration screen in the future.
For styling, a error screen and loading animation is introduced.

Signed-off-by: Luca Bretting <luca.bretting@fau.de>