feat: runnner of android processes
The commit implements both a server and a client process communicating over stdin and stdout.
Adb is the established way of launching and controlling processes on an android machine.
Stdin and Stdout is chosen, because thats the way to communicate over an adb shell.

The server process on android is started with a json request that specifies what process it should run:

```json
{
  "command": "/data/local/tmp/backend-daemon",
  "args": [],
  "env": { "RUST_LOG": "trace" },
  "root": true
}
```

It sends the logs over stdout in the following format:

{
  "Stdin": "...",
  "Stderr": "...",
  "InternalError": "..."
}

The server exits when the process stops, or when the client on the host sends `"Exit"` over stdin.

The client/host process requires a specification in the following format:

```json
{
  "runner_path": "./target/x86_64-android-linux/release/runner-server",
  "args": [],
  "env": { "RUST_LOG": "trace" },
  "root": true
}

Also, it takes a path to the executable as last argument, to be compatible with `cargo run` and `cargo test`.

The client connects to the android device via adb, pushes the executable and the server.
Afterwards it executes the server over adb with the appropriate request.

Signed-off-by: Felix Hilgers <felix.hilgers@fau.de>